流程操作记录工具类

+++ src/main/java/com/yunche/loan/domain/param/ApprovalParam.java
@@ -30,7 +30,7 @@ public class ApprovalParam {
     /**
      * 增补类型：  1-电审资料增补;  2-资料审核增补;
      */
-    private Integer supplementType;
+    private Byte supplementType;
     /**
      * 要求增补内容
      */
